requirements.txt
django==5.1.7
djangorestframework==3.15.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
